% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agz}
\alias{agz}
\title{Age group labels}
\format{
A character vector of length 17.
}
\usage{
agz
}
\description{
5-year age band labels used throughout the package (17 bands,
0-4 to 80+).
}
\keyword{datasets}
