% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterhandling.R
\name{create_demographic_parms}
\alias{create_demographic_parms}
\title{TODO}
\usage{
create_demographic_parms(
  tc = 1970:2020,
  nnat = 1,
  nrisk = 1,
  npost = 1,
  nstrain = 1,
  nprot = 1,
  migrationdata = list(),
  riskdata = list(),
  postdata = list(),
  straindata = list(),
  protdata = list(),
  verbose = FALSE
)
}
\arguments{
\item{tc}{TODO}

\item{nnat}{TODO}

\item{nrisk}{TODO}

\item{npost}{TODO}

\item{nstrain}{TODO}

\item{nprot}{TODO}

\item{migrationdata}{list of migration data see details}

\item{riskdata}{list of risk data see details}

\item{postdata}{TODO}

\item{straindata}{TODO}

\item{protdata}{TODO}

\item{verbose}{give more feedback}
}
\value{
list of parameters to run odin model
}
\description{
A short description...
}
\details{
Additional details...
}
\examples{
## example without nativity class used
pms <- create_demographic_parms() #create UK parameters
## version with 2 static nativity classes
pms <- create_demographic_parms(nnat = 2, migrationdata = list(propinitnat = c(0.9,0.1)))
## version with 2 static nativity classes and 2 static risk classes
pms <- create_demographic_parms(nnat = 2, nrisk = 2,
                                migrationdata = list(propinitnat = c(0.9,0.1)),
                                riskdata = list(propinitrisk = c(0.9,0.1)))
}
\author{
Pete Dodd
}
