% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterhandling.R
\name{known_parameters}
\alias{known_parameters}
\title{Print/test known parameters}
\usage{
known_parameters(parname = NULL, quiet = FALSE)
}
\arguments{
\item{parname}{if supplied consider only this parameter name; otherwise (default) print info on all parameters}

\item{quiet}{if TRUE (default) do not print info, only test whether parameter name is known}
}
\value{
no return value
}
\description{
Print info or test known parameters
}
\details{
This has no return value but can be used to either print info on known parameters (all, or specific parameters),
and will test for unknown parameters.

Info is returned organised by the parameter list they should be part of, and a short description and the expected
dimensions are printed.
}
\examples{
known_parameters()                     ##print all info
known_parameters("nonsense")           ##throws error as name not known
known_parameters("CDR_raw")            ##knows this parameter, prints info on it
known_parameters("CDR_raw",quiet=TRUE) ##knows this parameter, does not print info
}
\author{
Pete Dodd
}
