% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputhandling.R
\name{plt_DemoGrowth}
\alias{plt_DemoGrowth}
\title{TODO}
\usage{
plt_DemoGrowth(outdat)
}
\arguments{
\item{outdat}{}
}
\value{
a ggplot2 plot object
}
\description{
.. content for \description{} (no empty lines) ..TODO
}
\details{
.. content for \details{} ..TODO
}
\author{
Pete Dodd
}
