% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputhandling.R
\name{plt_DemoSnapshots}
\alias{plt_DemoSnapshots}
\title{Visualising Demographic snapshots}
\usage{
plt_DemoSnapshots(outdata)
}
\arguments{
\item{outdata}{a data.table returned by \code{runmodel} with \code{raw=FALSE}}
}
\value{
a \code{ggplot2} plot object
}
\description{
A short description...
}
\details{
Additional details...
}
\examples{
pms <- create_demographic_parms() #create UK parameters
out <- runmodel(pms)              #run model with these
out                               #inspect
## visualize
plt_DemoSnapshots(out)  
}
\seealso{
\code{\link[=plt_DemoGrowth]{plt_DemoGrowth()}}

\code{\link[=plt_TBSnapshots]{plt_TBSnapshots()}}
}
\author{
Pete Dodd
}
