% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputhandling.R
\name{plt_TB_rates}
\alias{plt_TB_rates}
\title{Title}
\usage{
plt_TB_rates(outdata, rate_type, by_layer)
}
\arguments{
\item{outdata}{model output}

\item{rate_type}{One of incidence, notification or mortality}

\item{by_layer}{one of natcat, risk, post, strain, prot}
}
\value{
ggplot of rates by age and sex stratified by target layer
}
\description{
Title
}
