% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputhandling.R
\name{plt_TBSnapshots}
\alias{plt_TBSnapshots}
\title{TODO}
\usage{
plt_TBSnapshots(outdat)
}
\arguments{
\item{outdat}{a data.table eg returned by runmodel with raw=FALSE}
}
\value{
ggplot2 plot
}
\description{
.. content for \description{} (no empty lines) ..TODO
}
\details{
.. content for \details{} ..TODO
}
