% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterhandling.R
\name{create_demographic_parms}
\alias{create_demographic_parms}
\title{TODO}
\usage{
create_demographic_parms(tc = 1970:2020, propinitnat = 1, verbose = FALSE)
}
\arguments{
\item{tc}{TODO}

\item{propinitnat}{vector of length nnat with initial proportions}

\item{verbose}{give more feedback}
}
\value{
list of parameters to run odin model
}
\description{
.. content for \description{} (no empty lines) ..TODO
}
\details{
.. content for \details{} ..TODO
}
\author{
Pete Dodd
}
