% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterhandling.R
\name{create_demographic_parms}
\alias{create_demographic_parms}
\title{TODO}
\usage{
create_demographic_parms(
  tc = 1970:2020,
  propinitnat = 1,
  propinitrisk = 1,
  birthrisk = 1,
  propinitpost = 1,
  propinitstrain = 1,
  propinitprot = 1,
  migrationdata,
  verbose = FALSE
)
}
\arguments{
\item{tc}{TODO}

\item{propinitnat}{vector of length nnat with initial proportions}

\item{propinitrisk}{vector of length nrisk with initial proportions}

\item{birthrisk}{where to births go in risk?}

\item{propinitpost}{initial post-TB proportions}

\item{propinitstrain}{initial TB strain proportions}

\item{propinitprot}{initial TB protection proportions}

\item{migrationdata}{list of migration data see details}

\item{verbose}{give more feedback}
}
\value{
list of parameters to run odin model
}
\description{
.. content for \description{} (no empty lines) ..TODO
}
\details{
.. content for \details{} ..TODO
}
\author{
Pete Dodd
}
