% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterhandling.R
\name{create_demographic_parms}
\alias{create_demographic_parms}
\title{TODO}
\usage{
create_demographic_parms(
  tc = 1970:2020,
  nnat = 1,
  nrisk = 1,
  npost = 1,
  nstrain = 1,
  nprot = 1,
  migrationdata = list(),
  riskdata = list(),
  postdata = list(),
  straindata = list(),
  protdata = list(),
  verbose = FALSE
)
}
\arguments{
\item{tc}{TODO}

\item{nnat}{TODO}

\item{nrisk}{TODO}

\item{npost}{TODO}

\item{nstrain}{TODO}

\item{nprot}{TODO}

\item{migrationdata}{list of migration data see details}

\item{riskdata}{list of risk data see details}

\item{postdata}{TODO}

\item{straindata}{TODO}

\item{protdata}{TODO}

\item{verbose}{give more feedback}
}
\value{
list of parameters to run odin model
}
\description{
.. content for \description{} (no empty lines) ..TODO
}
\details{
.. content for \details{} ..TODO
}
\author{
Pete Dodd
}
