% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parms.R
\name{create_parms}
\alias{create_parms}
\title{Creates parameters for model}
\usage{
create_parms(
  tc = 1970:2020,
  nnat = 1,
  nrisk = 1,
  npost = 1,
  nstrain = 1,
  nprot = 1,
  migrationdata = list(),
  riskdata = list(),
  postdata = list(),
  straindata = list(),
  protdata = list(),
  tbparms = list(),
  verbose = FALSE
)
}
\arguments{
\item{tc}{TODO}

\item{nnat}{TODO}

\item{nrisk}{TODO}

\item{npost}{TODO}

\item{nstrain}{TODO}

\item{nprot}{TODO}

\item{migrationdata}{list of migration data see details}

\item{riskdata}{list of risk data see details}

\item{postdata}{TODO}

\item{straindata}{TODO}

\item{protdata}{TODO}

\item{tbparms}{list of TB parameters; any not supplied take defaults (see \code{known_parameters()}). Time-varying entries (one row per \code{tc}) include \code{CDR_raw}, \code{migr_TBD_raw}, \code{migr_TBI_raw} and \code{betat_raw} (a length \code{length(tc)} multiplier on the force of infection, default all ones). \code{IRRnat} (length \code{nnat}, default all ones) multiplies progression from both latent states by nativity class. \code{progt_slow_raw} and \code{progt_fast_raw} (dimensions \code{length(tc)}, \code{nage}, \code{nnat}; default all ones) are time-varying multipliers on slow (reactivation) and fast progression by age and nativity class}

\item{verbose}{give more feedback}
}
\value{
list of parameter for model
}
\description{
Creates parameters for model
}
\author{
Pete Dodd
}
