% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterhandling.R
\name{create_parms}
\alias{create_parms}
\title{Creates parameters for model}
\usage{
create_parms(
  tc = 1970:2020,
  nnat = 1,
  nrisk = 1,
  npost = 1,
  nstrain = 1,
  nprot = 1,
  migrationdata = list(),
  riskdata = list(),
  postdata = list(),
  straindata = list(),
  protdata = list(),
  tbparms = list(),
  verbose = FALSE
)
}
\arguments{
\item{tc}{TODO}

\item{nnat}{TODO}

\item{nrisk}{TODO}

\item{npost}{TODO}

\item{nstrain}{TODO}

\item{nprot}{TODO}

\item{migrationdata}{list of migration data see details}

\item{riskdata}{list of risk data see details}

\item{postdata}{TODO}

\item{straindata}{TODO}

\item{protdata}{TODO}

\item{tbparms}{TODO}

\item{verbose}{give more feedback}
}
\value{
list of parameter for model
}
\description{
Creates parameters for model
}
\author{
Pete Dodd
}
