% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputhandling.R
\name{plt_DemoSnapshots}
\alias{plt_DemoSnapshots}
\title{TODO}
\usage{
plt_DemoSnapshots(outdat)
}
\arguments{
\item{outdat}{a data.table eg returned by runmodel with raw=FALSE}
}
\value{
a ggplot2 plot object
}
\description{
.. content for \description{} (no empty lines) ..TODO
}
\details{
.. content for \details{} ..TODO
}
\author{
Pete Dodd
}
