% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputhandling.R
\name{plt_TBSnapshots}
\alias{plt_TBSnapshots}
\title{Title}
\usage{
plt_TBSnapshots(outdata, by_layer)
}
\arguments{
\item{outdata}{modelled output}

\item{by_layer}{one of natcat, risk, post, strain, prot}
}
\value{
Population pyramid
}
\description{
Title
}
