% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputhandling.R
\name{plt_TBSnapshots}
\alias{plt_TBSnapshots}
\title{Visualising TB Snapshots}
\usage{
plt_TBSnapshots(outdat)
}
\arguments{
\item{outdat}{a data.table eg returned by \code{runmodel} with \code{raw=FALSE}}
}
\value{
\code{ggplot2} plot
}
\description{
A short description...
}
\details{
Additional details...
}
\examples{
##' pms <- create_demographic_parms() #create UK parameters
out <- runmodel(pms)              #run model with these
out                               #inspect
## visualize
plt_TBSnapshots(out)  
}
\seealso{
\code{\link[=plt_DemoSnapshots]{plt_DemoSnapshots()}}

\code{\link[=plt_DemoGrowth]{plt_DemoGrowth()}}
}
