% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_parameters.R
\name{print_helper}
\alias{print_helper}
\title{A helper function for \code{known_parameters}}
\usage{
print_helper(parname = NULL, parmlist)
}
\arguments{
\item{parname}{a specific parameter name (Default NULL)}

\item{parmlist}{one of the lists of parameters}
}
\value{
no return only printing
}
\description{
A helper function
}
\details{
See \code{known_parameters}
}
\author{
Pete Dodd
}
