% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelrunning.R
\name{runmodel}
\alias{runmodel}
\title{runmodel}
\usage{
runmodel(p, times, raw = FALSE)
}
\arguments{
\item{p}{parameter list}

\item{times}{vector of times to run the model over - if not given, will construct from p}

\item{raw}{return the raw matrix ODE output or a data.table? (default FALSE)}
}
\value{
a matrix of timeseries
}
\description{
A Simple Wrapper to Run Model
}
\details{
content for details here
}
\author{
Pete Dodd
}
